Add files via upload
The graph for the pseudoneutron are presented. The mass, volume, charge of the pseudoneutron are equal to the mass, charge, and volume of the neutron.
The result for the pseudoneutron is obtained by replacing the matrix a1 in the Program with the matrix for the pseudoneutron, pseudoproton:
a1 = array ([[2.0 , 2.0, 1.0, 1.0, 0.0, 0.0], [0.0, 0.0, 1.0, 0.0, 1.0, 0.0], 
                     [0.0, 0.0, 0.0, 0.0, 0.0, 1.0], [1.0, 0.0, 0.0, 2.0, 2.0, 1.0], 
                     [0.0, 1.0, 0.0, 0.0, 0.0, 1.0], [0.0, 0.0, 1.0, 0.0, 0.0, 0.0]])
A computation using machine learning will show the impossibility of obtaining a pseudoproton, since “Covariance of the parameters could not be estimated”, if only virtual. 
The pseudoneutron must be a short-lived particle.