Program version 2.0
Project name: Ver. 2.0. Using a python program to identify patterns inside a proton, neutron, pseudo neutron, virtual pseudo proton.

Description: The author's approach has been developed. It allowed writing the presented program in python, which in a couple of minutes on a laptop Lenovo (Intel(R) Core™ i7-3630QM CPU @ 2.40 GHz) calculates, visualizes data on 9 shells of a proton, neutron, pseudo neutron, virtual pseudo proton. Visualization made it possible to reveal the analytical relationship between charge, mass, volumes inside the particles.
As the initial data for the program, data were selected experimentally obtained by Nobel laureates and several recent published achievements of various scientists. The known data on quarks “u”, “d”, proton, neutron were used. All initial data and data obtained by the program are displayed in the form of separate tables. 
The program uses common python attributes, including machine learning.
The project made it possible to reveal that the relationship between charge, mass, the volume inside particles is described by several analytical 3D segments on straight lines, which are interconnected in triangles. For a proton, a neutron is two connected to each other on one side of the triangle. For a pseudo neutron, a virtual pseudo proton, these are three such triangles. This is clearly visualized in the graphs: “Interrelation of mass, volume, charge within a neutron Ver 2.0”; “Interrelation of mass, volume, charge within a proton Ver 2.0”; “Interrelation of mass, volume, charge within a pseudo proton Ver 2.0”; “Interrelation of mass, volume, charge within a pseudo neutron Ver 2.0”.
In general, for a proton, the equation for the first line segment will be as follows:
{█((m-pm1)/P1= (q-pq1)/P2= (v-pv1)/P3@ap1∈ [pm1;pm2]; [pq1;pq2]; [pv1;pv2] )┤
Where: 
pm1, pq1, pv1, pm2, pq2, pv2 - values of mass, charge, volume in accordance with the Project, respectively,
ap1 - designation of the first line segment, 
P(P1, P2, P3) - directing vector of a straight line segment.
The project simplified the understanding of the structure of the proton, neutron; predicted the presence of a new particle - a pseudo neutron. This knowledge simplifies the control of thermonuclear reactions.

Installation: To implement the Project, the third version of python is required. The project does not require any special skills to install. It will work on Jupiter Notebook. If you are missing some packages, then they are easily loaded using pip. 

Usage: To start the Project, it is enough to use Run. Everything will be done automatically. The program contains all the necessary comments. Therefore, if you need to use the received data in another program, you can easily find a place to insert the script. If you need to calculate, other elementary particles that are absent in the program, please contact the author.

Credits: The author of the project is Ivan Naborshchikov. Detailed information can be found on Linkedin: www.linkedin.com/in/ivan-naborshchikov