Version 3.1 introduced
Version 3.1 calculates and visualizes fundamental interactions. The amended code makes it possible to assert that any three fundamental interactions have the shape of a spatial triangle. From this, it follows that all four fundamental interactions represent a pyramid.  The added ability to use plot.ly allows the user to interactively view all the dependencies that can be obtained from the results obtained by the program.