Program Ver 3.1
Version 3.1 clearly demonstrates the relationship of mass, charge, volume, fundamental interactions based on a proton, neutron, pseudo proton, pseudo neutron.  Based on the program, we can talk about the unification of fundamental interactions in the form of a pyramid. The program allows us to go to the relationship of mass, charge, volume with fundamental interactions as 7D space.  Thus, we can optimize technological processes and cut costs. 

Version 3.1 allows the user to apply plot.ly for all values obtained by the program for proton, neutron, pseudo proton, pseudo neutron. Thus, the user can interactively create 2D and 3D interactive graphs, charts, histograms of interest to him. Version 3.1 fixes a bug that was in version 3.0 .; indexes in "gravity" are affixed.