Ver 3.2 With additional explanation 1
Ver 3.2 Schedule: "The relationship between frequency, gravity, and electromagnetism, neutron, and proton. "
The frequency was calculated taking into account the speed of light and Compton wavelength for the shells of the proton, neutron.  The project ver3.1 contains all the necessary calculations made on the basis of generally accepted formulas, constants in the Python3 language.